Changed naming convention, started normal_form
instead of being of the form `longname` names are of the form `longname
and break when reaching a character that is not valid for a name (such
as white space). see notes.txt and source for which characters indicate
the end of a name

normal_form reduces expressions to a regular form that can be compared
to determine if two lambdas/expressions are equivalent. Work in progress